Values without decimal point or exponent should be int.
This fixes current issue where only 0 is an integer,
and all other numbers are doubles.

+++ lib/src/parse.dart
@@ -468,7 +468,7 @@ Map<String, Token? Function()> lexStates = {
       return null; //$
     }
 
-    return newToken('numeric', sign * double.parse(buffer!));
+    return newToken('numeric', sign * int.parse(buffer!));
   },
   'decimalPointLeading': () {
     if (util.isDigit(c)) {
